requirements.txt
bmipy==2.0.1
cftime==1.6.3
click==8.1.7
cython==3.0.10
entrypoints==0.4
h5netcdf==1.3.0
h5py==3.11.0
loguru==0.7.2
matplotlib==3.8.4
numpy==1.26.4
pandas==2.2.2
pyyaml==6.0.1
scipy==1.13.1
seaborn==0.13.2
tqdm==4.66.4
